LayoutTests/fast/viewport/viewport-60.html

<head>
    <title>Check that maximum-scale=device-height is 10.</title>
    <meta name="viewport" content="maximum-scale=device-height, initial-scale=10">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
