Safari-611.1.10/LayoutTests/fast/viewport/viewport-60-expected.txt

ALERT: viewport size 32x35.2 scale 10 with limits [10, 10] and userScalable true
